fix(traverse): Avoid scale/subobject name clash
Adds a test to reproduce the issues when a `plone.app.imaging` (and thus also
`eea.depiction`) scale clashes with an actual ZODB OFS subobject.  Also includes a fix
that mimics the fix in [my `plone.app.imaging`
PR](https://github.com/plone/plone.app.imaging/pull/41).

Because the same logic needs to be used in both the `plone.app.imaging` traverser and
the `eea.depiction` traverser, this fix should work and resolve the issue even before or
the `plone.app.imaging` PR is merged or if it never is.  IOW, we can treat these
separately.